clean up  backend-mock a bit
